kernel: selinux/sepolicy: handle selinux policydb backports
this handles selinux policydb handling backports which is becoming a thing on 4.x
and will probably become more and more common as time goes on.

we just differentiate this on policydb.h

5.4+:	struct type_datum **type_val_to_struct;
5.1: 	struct type_datum **type_val_to_struct_array;
<5.1:	struct flex_array *type_val_to_struct_array;

as for ksu_kvrealloc, we dont have unsafe_memcpy on older kernels.
its just a macro of __builtin_memcpy anyway.
we do this to assure kasan/fortify_source won't cry.

ref:
https://github.com/tiann/KernelSU/pull/3087
https://github.com/tiann/KernelSU/pull/3089
https://github.com/tiann/KernelSU/commit/438559c82484214cf0317df448558825e6ab137e *

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>